requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
twilio==9.2.3
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
